feature/add_new_customer

+++ internal/entity/customer_response_dto.go
@@ -4,7 +4,7 @@ package entity
 
 // Customer -.
 type CustomerResponseDTO struct {
-	CustomerId  uint32 `json:"customerId"`
+	CustomerId  uint64 `json:"customerId"`
 	FirstName   string `json:"firstName"`
 	LastName    string `json:"lastName"`
 	Email       string `json:"email"`
